+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/jobs/DelayJob.java
@@ -121,6 +121,11 @@ public class DelayJob extends AutomationJob {
     }
 
     @Override
+    public void stop() {
+        endJob = true;
+    }
+
+    @Override
     public String getType() {
         return JOB_NAME;
     }